Added secio check back to p2p
go-libp2p replaced the NoEncryption method with the NoSecurity variable. Since the p2p tutorial still uses the secio parameter and flag, it makes sense to showcase the original. Otherwise the code becomes confusing since the secio parameter is not used anywhere, to understand this change it is necessary to browse through the closed issues, which is not that user-friendly. The alternative is to remove the secio flag and parameter all together, but this would require the blog post to be updated, while this change could also cause an update, the update is minimal.